+++ src/main/java/org/dma/gbdt4spark/algo/gbdt/histogram/GradPair.java
@@ -11,7 +11,7 @@ public interface GradPair {
 
     GradPair subtract(GradPair gradPair);
 
-    void timesBy(float x);
+    void timesBy(double x);
 
     float calcGain(GBDTParam param);
